--- a/tox.ini
+++ b/tox.ini
@@ -8,6 +8,7 @@ envlist=
 
 [testenv]
 commands= py.test --lsof -rfsxX {posargs:testing}
+passenv = USER USERNAME
 deps=
     nose
     mock