Fix for #105 

Using a sub-directory as root in the `TEMP` directory speeds up `make_numbered_dir` calls, because it will loop over all files to find an unused directory name.

```python
import pytest

pytest_plugins = 'pytester'
@pytest.mark.parametrize('i', range(10))
def test_foo(i, testdir):
    print(str(testdir.tmpdir))  
```

**before**:

```
========================== slowest 3 test durations ===========================
0.46s setup    test_foo.py::test_foo[0]
0.00s setup    test_foo.py::test_foo[7]
0.00s teardown test_foo.py::test_foo[5]
```

**after**:

```
========================== slowest 3 test durations ===========================
0.00s setup    test_foo.py::test_foo[0]
0.00s setup    test_foo.py::test_foo[7]
0.00s teardown test_foo.py::test_foo[5]
```

The speed-up depends on the number of files in the user's `$TEMP` directory.