@The-Compiler 2015-07-25
I think this change somehow broke things for me, but I can't reproduce it.

I had a `/tmp/pytest-florian` which is a broken symlink (the "broken" thing is what I can't reproduce):

```
lrwxrwxrwx 1 florian florian 36 Jul 24 12:35 /tmp/pytest-florian -> /tmp/pytest-62/test_exception_exc1_0
```

(so `/tmp/pytest-62/test_exception_exc1_0` doesn't exist)

When running the current master, I got an error with every test using `tmpdir`:

```
self = <py._error.ErrorMaker object at 0x7f9bd78b89b0>, func = <built-in function mkdir>, args = ('/tmp/pytest-florian',), kwargs = {}, __tracebackhide__ = False, cls = <class 'py.error.EEXIST'>
value = FileExistsError(17, 'File exists'), tb = <traceback object at 0x7f9bd5090848>, errno = 17

    def checked_call(self, func, *args, **kwargs):
        """ call a function and raise an errno-exception if applicable. """
        __tracebackhide__ = True
        try:
            return func(*args, **kwargs)
        except self.Error:
            raise
        except (OSError, EnvironmentError):
            cls, value, tb = sys.exc_info()
            if not hasattr(value, 'errno'):
                raise
            __tracebackhide__ = False
            errno = value.errno
            try:
                if not isinstance(value, WindowsError):
                    raise NameError
            except NameError:
                # we are not on Windows, or we got a proper OSError
                cls = self._geterrnoclass(errno)
            else:
                try:
                    cls = self._geterrnoclass(_winerrnomap[errno])
                except KeyError:
                    raise value
>           raise cls("%s%r" % (func.__name__, args))
E           py.error.EEXIST: [File exists]: mkdir('/tmp/pytest-florian',)
```

Removing the broken symlink fixed it, but I wonder how that happened...